renames the util file for user
